fixed startup/shutdown with specific features

+++ arangod/Scheduler/SchedulerFeature.cpp
@@ -273,7 +273,7 @@ class HangupTask : public SignalTask {
 #endif
 
 void SchedulerFeature::buildScheduler() {
-  _scheduler = new SchedulerLibev(_nrSchedulerThreads, _backend);
+  _scheduler = new SchedulerLibev(_nrSchedulerThreads, static_cast<int>(_backend));
   SCHEDULER = _scheduler;
 }
